STRD-17

Approved-by: Vinicius Bettiol <vinicius.bettiol@canoadigital.com>
Approved-by: Gil Tavares <gil.tavares@canoadigital.com>
Approved-by: Rayane de Araujo Nunes <rayane.araujo@canoadigital.com>
Approved-by: Thomas Edson Nunes <thomas.nunes@canoadigital.com>

+++ Jobs/Asterisk.php
@@ -2,7 +2,6 @@
 
 include "IAsterisk.php";
 include "DialogFlow.php";
-
 require __DIR__ . '/../vendor/autoload.php';
 
 use \Curl\Curl;
@@ -17,7 +16,7 @@ class Asterisk implements IAsterisk
 
     //config with default vars
     private $min_confidence = 0.80;
-    private $message_not_understand = 'Desculpe, não consegui te endender, poderia repetir?';
+    private $message_not_understand = 'not_understand';
 
     /**
      * asterisk constructor.
@@ -49,6 +48,67 @@ class Asterisk implements IAsterisk
     {
 
         $this->curl->close();
+
+    }
+
+    /**
+     * create all tables and fill it with the default value
+     */
+    public function setupDB()
+    {
+
+        //put here all table files, like migrations on Laravel
+        require_once "database/AllDefaultMessages.php";
+
+        //fill default values
+        $defaults[] = array('textName'  => 'MENS_AGUARDE',
+                            'textValue' => 'Certo, Aguarde só um momentinho que vou verificar');
+        $defaults[] = array('textName'  => 'MENS_NAO_ENTENDI',
+                            'textValue' => 'Nao entendi sua pergunta, poderia repetir?');
+        $defaults[] = array('textName'  => 'MENS_NAO_CONSEGUI_AJUDAR',
+                            'textValue' => 'Infelizmente nao conseguirei te ajudar, entre em contato em horario comercial ou acesse nosso site www.astrocentro.com.br, obrigado');
+        $defaults[] = array('textName'  => 'MENS_DUVIDA',
+                            'textValue' => 'Consegui tirar sua dúvida, diga sim ou não ');
+        $defaults[] = array('textName'  => 'MENS_CONSEGUI_AJUDAR',
+                            'textValue' => 'Fico muito feliz em ter ajudado, caso tenha novas dúvidas nos contate, ficaremos felizes em te atender. Forte abraço!!');
+        $defaults[] = array('textName'  => 'MENS_ENTENDI',
+                            'textValue' => 'Entendi, vamos tentar de outra forma então, me diga em poucas palavras o que você precisa');
+
+        //store on DB
+        foreach ($defaults as $default)  AllDefaultMessages::Create($default);
+
+    }
+
+    /**
+     * set all default messages from DB
+     */
+    public function getDefaultMessages()
+    {
+
+        require_once __DIR__ . "/../bootstrap.php";
+
+        $this->agi->exec("NOOP", "getDefaultMessages\ ");
+
+        $time_start = microtime(true);
+
+        foreach (AllDefaultMessages::All() as $message) {
+
+            //translate text to audio
+            $ret = $this->textToSpeech( $message->textValue );
+
+            $ret['localFile'] = $this->convertFileToAsterisk($ret['transcript'], $ret['fileName']);
+
+            $this->agi->exec("NOOP",  $message->textName . "\ ". $ret['localFile']);
+
+            echo "\n";
+
+            $this->agi->set_variable($message->textName, $ret['localFile'] );
+
+        }
+
+        $time_end = microtime(true);
+
+        $this->agi->exec("NOOP", "Total\ Execution\ Time\ getDefaultMessages:\ " .  (($time_end - $time_start)) );
 
     }
 
@@ -90,9 +150,21 @@ class Asterisk implements IAsterisk
             $time_start = microtime(true);
 
 
-            //to avoid null answers
-            if(!$astrid_answer)
-                $astrid_answer = $this->message_not_understand;
+            echo "\n";
+
+            //to avoid not understand answers
+            if($astrid_answer == 'Desculpe, mas não consegui te entender' || !$astrid_answer){
+
+                $this->agi->set_variable("not_understand", 1 );
+                return 1;
+
+            }else{
+
+                $this->agi->set_variable("not_understand", 0 );
+
+            }
+
+            echo "\n";
 
             //translate text to audio
             $ret = $this->textToSpeech( $astrid_answer );
@@ -103,7 +175,8 @@ class Asterisk implements IAsterisk
 
         }else{
 
-            $ret = $this->textToSpeech( $this->message_not_understand );
+           $this->agi->set_variable("not_understand", 1 );
+           return 1;
 
         }
 
@@ -183,12 +256,19 @@ class Asterisk implements IAsterisk
         //translate text to audio
         $ret = $this->textToSpeech( $message );
 
+
+ 	    $this->agi->exec("NOOP", "extTextToSpeech\ " . $message );
+        $this->agi->exec("NOOP", "extTextToSpeech\ " . $ret['transcript'] );
+
+
         $time_end = microtime(true);
 
         $this->agi->exec("NOOP", "Total\ Execution\ Time\ textToSpeech:\ " .  (($time_end - $time_start)) );
 
 
         $ret['localFile'] = $this->convertFileToAsterisk($ret['transcript'], $ret['fileName']);
+
+	    $this->agi->exec("NOOP", "extTextToSpeech\ ". $ret['localFile']);
 
         echo "\n";
 
@@ -266,12 +346,14 @@ class Asterisk implements IAsterisk
     public function textToSpeech($message)
     {
 
+        $this->agi->exec("NOOP", "textToSpeech\ " . $message);
+
         $this->curl->get( getenv("TRANSLATE-API-URL") . '/text-to-speech', array(
                                                                             "message" => $message,       
                                                                            ));
         if ($this->curl->error) {
             
-            $ret['transcript'] = 'Error: ' . $this->curl->errorCode . ': ' . $this->curl->errorMessage . "\n";
+            $ret['transcript'] = 'Error:\ ' . $this->curl->errorCode . '\: ' . $this->curl->errorMessage . "\ \n";
             $ret['status'] = 0;
 
         }else{
@@ -280,7 +362,7 @@ class Asterisk implements IAsterisk
             $ret['fileName'] =  substr($ret['transcript'], strrpos($ret['transcript'], '/')+1);
             $ret['status'] = 1;
         }
-        
+
         return $ret;
 
     }
@@ -338,14 +420,17 @@ class Asterisk implements IAsterisk
     public function convertFileToAsterisk($s3Url, $fileName)
     {
 
-        $audio = file_get_contents($s3Url);
-
-        //save file on /tmp
-        file_put_contents("/tmp/" . $fileName, $audio);
-
         $FileNameWithOutExt = substr($fileName, 0, strpos($fileName, '.'));
 
-        system("lame --decode /tmp/$fileName - | sox -v 0.5 -t wav - -t wav -b 16 -r 8000 -c 1 $FileNameWithOutExt.wav");
+        if(!file_exists("/tmp/" . $FileNameWithOutExt . ".wav" )){
+
+            $audio = file_get_contents($s3Url);
+
+            //save file on /tmp
+            file_put_contents("/tmp/" . $fileName, $audio);
+
+            system("lame --decode /tmp/$fileName - | sox -v 0.5 -t wav - -t wav -b 16 -r 8000 -c 1 /tmp/$FileNameWithOutExt.wav");
+        }
 
         return "/tmp/" . $FileNameWithOutExt;
